+++ Library/Keyboard/Keyboard.c
@@ -1,7 +1,7 @@
 #include "keyboard.h"
 extern xSemaphoreHandle Keyboard_semapfore;
 
-extern int Display_update;
+extern uint8_t Display_update;
 extern xSemaphoreHandle Display_semaphore;
 extern char Keyboard_press_code;
